requirements.txt
asgiref==3.6.0
black==23.3.0
boto3==1.26.115
botocore==1.29.115
cfgv==3.3.1
click==8.1.3
coverage==7.2.5
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-storages==1.13.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
filelock==3.12.0
flake8==6.0.0
flake8-django==1.2
identify==2.5.24
iniconfig==2.0.0
isort==5.12.0
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
pre-commit==3.3.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
virtualenv==20.23.0
